Move dtos to new package
