+++ feature-login/src/commonMain/kotlin/data/Login.kt
@@ -5,7 +5,7 @@
 package org.pointyware.xyz.feature.login.data
 
 import org.pointyware.xyz.core.entities.profile.Profile
-import org.pointyware.xyz.core.data.dtos.Authorization
+import org.pointyware.xyz.api.dtos.Authorization
 
 /**
  * Combines the general authorization data with the user profile.
